24/3 Changes
- Tuned Angler and Launcher

- Adjusted shot map

- Adjusted paths for no Yoshi

- Decoupled Launcher and Angler
