+++ .pathplanner/settings.json
@@ -15,8 +15,8 @@
     "Debugging",
     "TopAutos"
   ],
-  "defaultMaxVel": 4.0,
-  "defaultMaxAccel": 3.5,
+  "defaultMaxVel": 2.5,
+  "defaultMaxAccel": 2.0,
   "defaultMaxAngVel": 540.0,
   "defaultMaxAngAccel": 720.0,
   "maxModuleSpeed": 4.5
