+++ src/main/deploy/pathplanner/autos/A2-W321.auto
@@ -14,13 +14,13 @@
         {
           "type": "named",
           "data": {
-            "name": "Shoot"
+            "name": "SpeakerShot"
           }
         },
         {
           "type": "named",
           "data": {
-            "name": "DeployYoshi"
+            "name": "IntakeOnly"
           }
         },
         {
