+++ src/main/deploy/pathplanner/paths/A2-W3.path
@@ -16,12 +16,12 @@
     },
     {
       "anchor": {
-        "x": 2.3336313575910204,
-        "y": 4.44
+        "x": 2.471490794833442,
+        "y": 4.35744750656168
       },
       "prevControl": {
-        "x": 2.1660946959870846,
-        "y": 4.720163335623921
+        "x": 2.303954133229506,
+        "y": 4.637610842185601
       },
       "nextControl": null,
       "isLocked": false,
@@ -32,8 +32,8 @@
   "constraintZones": [],
   "eventMarkers": [],
   "globalConstraints": {
-    "maxVelocity": 2.0,
-    "maxAcceleration": 1.0,
+    "maxVelocity": 2.5,
+    "maxAcceleration": 2.0,
     "maxAngularVelocity": 540.0,
     "maxAngularAcceleration": 720.0
   },
@@ -48,5 +48,5 @@
     "rotation": 0.0,
     "velocity": 0
   },
-  "useDefaultConstraints": false
+  "useDefaultConstraints": true
 }
