+++ src/main/java/frc/robot/subsystems/drive/GyroIOPigeon2.java
@@ -26,7 +26,7 @@ public class GyroIOPigeon2 implements GyroIO {
     gyro.getConfigurator().apply(new Pigeon2Configuration());
     gyro.getConfigurator().setYaw(0.0);
 
-    yaw.setUpdateFrequency(100.0);
+    yaw.setUpdateFrequency(50.0);
     yawVelocity.setUpdateFrequency(50.0);
 
     gyro.optimizeBusUtilization();
