+++ src/main/java/frc/robot/subsystems/drive/ModuleIOSparkMax.java
@@ -4,6 +4,7 @@
 
 package frc.robot.subsystems.drive;
 
+import com.ctre.phoenix6.BaseStatusSignal;
 import com.ctre.phoenix6.hardware.CANcoder;
 import com.revrobotics.CANSparkBase.ControlType;
 import com.revrobotics.CANSparkBase.IdleMode;
@@ -150,6 +151,11 @@ public class ModuleIOSparkMax implements ModuleIO {
 
     driveMotor.burnFlash();
     azimuthMotor.burnFlash();
+
+    BaseStatusSignal.setUpdateFrequencyForAll(
+  50, angleEncoder.getAbsolutePosition());
+
+    angleEncoder.optimizeBusUtilization();
   }
 
   @Override
