+++ src/main/java/frc/robot/subsystems/shooter/angler/AnglerIOSparkMax.java
@@ -21,7 +21,7 @@ public class AnglerIOSparkMax implements AnglerIO {
   private DutyCycleEncoder absoluteEncoder = new DutyCycleEncoder(0);
   private Encoder relativeEncoder = new Encoder(5, 6);
 
-  private Rotation2d absoluteEncoderOffset = Rotation2d.fromDegrees(179.0);
+  private Rotation2d absoluteEncoderOffset = Rotation2d.fromDegrees(179.0 + 1.43);
 
   private double appliedVolts = 0.0;
